requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
email-validator==2.2.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41
Werkzeug==3.1.3
gunicorn==23.0.0
matplotlib==3.10.3
numpy==2.3.1
Pillow==11.3.0
python-dotenv==1.0.0 